Create action model.
An action represents a single action in a game. The goal is to be able
to reproduce any step of the game by replaying the actions leading up to
that point in time.
